refactor(recommendations): move recommendation service files into a dedicated folder
