+++ src/hooks/useRecommendations.js
@@ -1,7 +1,7 @@
 // useRecommendations.js
 
 import { useState } from 'react';
-import recommendationService from '../services/recommendation.service';
+import recommendationService from '../services/recommendation/recommendation.service';
 
 function useRecommendations(products) {
   const [recommendations, setRecommendations] = useState([]);
